big brain forgot to change occurrences of vexer in the tests
